+++ src/blaze/system.clj
@@ -92,9 +92,9 @@
     (into {} (map (fn [[k v]] [k (resolvef k v)])) (ig/find-derived config key))))
 
 (def ^:private root-config
-  {:blaze/version "0.26.0"
+  {:blaze/version "0.26.1"
 
-   :blaze/release-date "2024-04-16"
+   :blaze/release-date "2024-04-28"
 
    :blaze/clock {}